# improve allocation/performance of ResolutionList