PR 2 extracted out of #40

Focused on ResolutionList
-  Avoid allocating a ResolutinoList if not needed
-  Improve performance of the Contains method

Performance measurements (purely this PR):
From ResolveFromType
|  Method |     Mean |   Error |  StdDev |      Min |      Max |   Median | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|--------:|--------:|---------:|---------:|---------:|-----:|-------:|------:|------:|----------:|
| Current | 223.8 ns | 1.26 ns | 1.18 ns | 221.6 ns | 225.7 ns | 223.6 ns |    1 | 0.0627 |     - |     - |     296 B |
| Master  | 226.5 ns | 1.43 ns | 1.34 ns | 224.5 ns | 229.2 ns | 227.1 ns |    1 | 0.0730 |     - |     - |     344 B |

From ResolveFromTypeWithInnerTypes 
|  Method |     Mean |   Error |  StdDev |      Min |      Max |   Median | Rank |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |---------:|--------:|--------:|---------:|---------:|---------:|-----:|-------:|------:|------:|----------:|
| Current | 319.3 ns | 2.11 ns | 1.97 ns | 316.2 ns | 322.2 ns | 319.6 ns |    1 | 0.0896 |     - |     - |     424 B |
| Master  | 325.4 ns | 2.31 ns | 2.16 ns | 322.5 ns | 328.7 ns | 324.9 ns |    1 | 0.1001 |     - |     - |     472 B |