@bollhals 2023-03-23
@SabotageAndi or @tzongithub There are still a whole lot of PR's open for the Bodi, They should be fairly simple to review and it would improve the performance for it. Can we align on how to proceed with it? It's been years already since I filed them :)