@SabotageAndi 2023-03-24
@bollhals Sorry that we never got to these PRs.
I left Tricentis last month, so I no longer work on SpecFlow.