@SabotageAndi 2023-03-25
@gasparnagy should still work for Tricentis as a freelancer.