@gasparnagy 2023-03-27
> @gasparnagy should still work for Tricentis as a freelancer.

That is unfortunately not true. They haven't renewed my contract (yet?). So @bollhals I suggest contacting them at info@specflow.org.